fix: Refactor hierarchy utils into questionset and content packages
