+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/utils/BranchingUtil.scala
@@ -2,6 +2,7 @@ package org.sunbird.utils
 
 import org.apache.commons.lang3.StringUtils
 import org.sunbird.common.dto.Request
+import org.sunbird.utils.questionset.HierarchyConstants
 import java.util
 import scala.jdk.CollectionConverters._
