+++ assessment-api/qs-hierarchy-manager/src/main/scala/org/sunbird/utils/HierarchyErrorCodes.scala
@@ -1,4 +1,4 @@
-package org.sunbird.utils
+package org.sunbird.utils.questionset
 
 object HierarchyErrorCodes {
     val ERR_INVALID_ROOT_ID: String = "ERR_INVALID_ROOT_ID"
