+++ assessment-api/qs-hierarchy-manager/src/test/scala/org/sunbird/managers/UpdateHierarchyManagerTest.scala
@@ -6,7 +6,7 @@ import org.parboiled.common.StringUtils
 import org.sunbird.common.JsonUtils
 import org.sunbird.common.dto.Request
 import org.sunbird.graph.OntologyEngineContext
-import org.sunbird.utils.HierarchyConstants
+import org.sunbird.utils.questionset.HierarchyConstants
 
 import scala.concurrent.ExecutionContext
