+++ content-api/content-actors/src/main/scala/org/sunbird/content/actors/CollectionActor.scala
@@ -5,7 +5,7 @@ import org.sunbird.actor.core.BaseActor
 import org.sunbird.common.dto.{Request, Response}
 import org.sunbird.graph.OntologyEngineContext
 import org.sunbird.managers.content.{HierarchyManager, UpdateHierarchyManager}
-import org.sunbird.utils.HierarchyConstants
+import org.sunbird.utils.content.HierarchyConstants
 
 import scala.concurrent.{ExecutionContext, Future}
