+++ content-api/hierarchy-manager/src/main/scala/org/sunbird/utils/HierarchyBackwardCompatibilityUtil.scala
@@ -1,4 +1,4 @@
-package org.sunbird.utils
+package org.sunbird.utils.content
 
 import java.util
