+++ content-api/hierarchy-manager/src/test/scala/org/sunbird/managers/TestUpdateHierarchy.scala
@@ -8,7 +8,7 @@ import org.sunbird.common.JsonUtils
 import org.sunbird.common.dto.Request
 import org.sunbird.common.exception.{ClientException, ResourceNotFoundException}
 import org.sunbird.graph.OntologyEngineContext
-import org.sunbird.utils.HierarchyConstants
+import org.sunbird.utils.content.HierarchyConstants
 
 class TestUpdateHierarchy extends BaseSpec {
